requirements.txt
Django==1.9.4
psycopg2==2.6.1
python-dateutil==2.5.0
six==1.10.0
tqdm==3.8.0